requirements.txt
strands-agents
bedrock-agentcore
boto3>=1.34.0
botocore[crt]
python-dotenv>=1.0.0
streamlit>=1.30
